+++ fs/loopback_linux.go
@@ -9,7 +9,6 @@ package fs
 
 import (
 	"context"
-	"path/filepath"
 	"syscall"
 
 	"golang.org/x/sys/unix"
@@ -41,41 +40,6 @@ var _ = (NodeListxattrer)((*LoopbackNode)(nil))
 func (n *LoopbackNode) Listxattr(ctx context.Context, dest []byte) (uint32, syscall.Errno) {
 	sz, err := unix.Llistxattr(n.path(), dest)
 	return uint32(sz), ToErrno(err)
-}
-
-func (n *LoopbackNode) renameExchange(name string, newparent InodeEmbedder, newName string) syscall.Errno {
-	fd1, err := syscall.Open(n.path(), syscall.O_DIRECTORY, 0)
-	if err != nil {
-		return ToErrno(err)
-	}
-	defer syscall.Close(fd1)
-	p2 := filepath.Join(n.RootData.Path, newparent.EmbeddedInode().Path(nil))
-	fd2, err := syscall.Open(p2, syscall.O_DIRECTORY, 0)
-	defer syscall.Close(fd2)
-	if err != nil {
-		return ToErrno(err)
-	}
-
-	var st syscall.Stat_t
-	if err := syscall.Fstat(fd1, &st); err != nil {
-		return ToErrno(err)
-	}
-
-	// Double check that nodes didn't change from under us.
-	inode := &n.Inode
-	if inode.Root() != inode && inode.StableAttr().Ino != n.RootData.idFromStat(&st).Ino {
-		return syscall.EBUSY
-	}
-	if err := syscall.Fstat(fd2, &st); err != nil {
-		return ToErrno(err)
-	}
-
-	newinode := newparent.EmbeddedInode()
-	if newinode.Root() != newinode && newinode.StableAttr().Ino != n.RootData.idFromStat(&st).Ino {
-		return syscall.EBUSY
-	}
-
-	return ToErrno(unix.Renameat2(fd1, name, fd2, newName, unix.RENAME_EXCHANGE))
 }
 
 var _ = (NodeCopyFileRanger)((*LoopbackNode)(nil))
