+++ fuse/request_darwin.go
@@ -9,5 +9,5 @@ const outputHeaderSize = 200
 const (
 	_FUSE_KERNEL_VERSION   = 7
 	_MINIMUM_MINOR_VERSION = 12
-	_OUR_MINOR_VERSION     = 12
+	_OUR_MINOR_VERSION     = 19
 )
